Remove EntityDatabaseObserverWrapper
