+++ Tests/StreamChatTests/Controllers/MemberController/MemberController_Tests.swift
@@ -488,7 +488,7 @@ final class MemberController_Tests: XCTestCase {
 private class TestEnvironment {
     @Atomic var memberUpdater: ChannelMemberUpdater_Mock?
     @Atomic var memberListUpdater: ChannelMemberListUpdater_Mock?
-    @Atomic var memberObserver: EntityDatabaseObserverWrapper_Mock<ChatChannelMember, MemberDTO>?
+    @Atomic var memberObserver: BackgroundEntityDatabaseObserver_Mock<ChatChannelMember, MemberDTO>?
     @Atomic var memberObserverSynchronizeError: Error?
 
     lazy var environment: ChatChannelMemberController.Environment = .init(
